requirements.txt
dynaconf
flickrapi
jinja2
requests
python-slugify
